Reduce zarr storage: float16 + zstd clevel=9 bitshuffle
- Save zarr as float16 instead of float32 (50% size reduction)
- Use Blosc zstd clevel=9 with bitshuffle (further compression)
- Dataset loads float16 zarr and casts to float32 automatically

+++ data/dataset.py
@@ -125,7 +125,7 @@ class BiosignalDataset(Dataset[BiosignalSample]):
         if path.endswith(".zarr"):
             import zarr
             arr = zarr.open(path, mode="r")
-            return torch.from_numpy(arr[:])
+            return torch.from_numpy(arr[:]).float()  # float16 zarr → float32
         return torch.load(path, weights_only=True, mmap=self._use_mmap)
 
     def __getstate__(self) -> dict:
